Updates tests to pester 5

+++ .build.ps1
@@ -8,14 +8,10 @@ param (
 
 Set-Alias MSBuild (Resolve-MSBuild) -ErrorAction SilentlyContinue
 
-task default Setup,
-             Build,
-             Test,
-             Publish
-
-task Setup SetBuildInfo,
-           InstallRequiredModules,
-           UpdateAppVeyorVersion
+task default Build,
+             Publish,
+             Setup,
+             Test
 
 task Build Clean,
            TestSyntax,
@@ -24,6 +20,12 @@ task Build Clean,
            UpdateMetadata,
            UpdateMarkdownHelp
 
+task Publish PublishToPSGallery
+
+task Setup SetBuildInfo,
+           UpdateRequiredModules,
+           UpdateAppVeyorVersion
+
 task Test TestModuleImport,
           TestAttributeSyntax,
           PSScriptAnalyzer,
@@ -31,8 +33,6 @@ task Test TestModuleImport,
           AddAppveyorCompilationMessage,
           UploadAppVeyorTestResults,
           ValidateTestResults
-
-task Publish PublishToPSGallery
 
 function GetBuildSystem {
     [OutputType([String])]
@@ -877,7 +877,7 @@ task SetBuildInfo -If (
     }
 }
 
-task InstallRequiredModules {
+task UpdateRequiredModules {
     # Installs the modules required to execute the tasks in this script into current user scope.
 
     $erroractionpreference = 'Stop'
@@ -892,10 +892,10 @@ task InstallRequiredModules {
                 Target    = 'CurrentUser'
             }
 
-            Configuration    = 'latest'
-            Pester           = 'latest'
-            PlatyPS          = 'latest'
-            PSScriptAnalyzer = 'latest'
+            Configuration    = '1.3.1'
+            Pester           = '5.1.0'
+            PlatyPS          = '0.14.0'
+            PSScriptAnalyzer = '1.19.1'
         }
     } catch {
         throw
@@ -905,16 +905,18 @@ task InstallRequiredModules {
 task UpdateAppVeyorVersion -If (
 Test-Path (Join-Path $buildInfo.Path.ProjectRoot 'appveyor.yml')
 ) {
-    $versionString = '{0}.{1}.{2}.{{build}}' -f @(
-        $buildInfo.Version.Major
-        $buildInfo.Version.Minor
-        $buildInfo.Version.Build
-    )
+    if (Join-Path $buildInfo.Path.ProjectRoot 'appveyor.yml' | Test-Path) {
+        $versionString = '{0}.{1}.{2}.{{build}}' -f @(
+            $buildInfo.Version.Major
+            $buildInfo.Version.Minor
+            $buildInfo.Version.Build
+        )
 
-    $path = Join-Path $buildInfo.Path.ProjectRoot 'appveyor.yml'
-    $content = Get-Content $path -Raw
-    $content = $content -replace 'version: .+', ('version: {0}' -f $versionString)
-    Set-Content $path -Value $content -NoNewLine
+        $path = Join-Path $buildInfo.Path.ProjectRoot 'appveyor.yml'
+        $content = Get-Content $path -Raw
+        $content = $content -replace 'version: .+', ('version: {0}' -f $versionString)
+        Set-Content $path -Value $content -NoNewLine
+    }
 }
 
 task Clean {
@@ -1399,30 +1401,27 @@ task TestModule {
             }
         }
 
-        # Prevent the default code coverage report appearing.
-        Import-Module Pester
-        & (Get-Module pester) {
-            $definition = Get-Content function:Write-CoverageReport
-            $definition = $definition -replace '(\$report.+Format-Table)', '# $1'
-            Set-Item function:Write-CoverageReport -Value $definition
-        }
-
-        Import-Module $buildInfo.Path.Build.Manifest -Global -ErrorAction Stop
-        $params = @{
-            Script     = @{
-                Path       = $path
-                Parameters = @{
-                    UseExisting = $true
-                }
+        Import-Module $buildInfo.Path.Build.Manifest -Global -ErrorAction Stop -Force
+        $configuration = @{
+            Run          = @{
+                Path     = $path
+                PassThru = $true
             }
-            OutputFile = Join-Path $buildInfo.Path.Build.Output ('{0}-nunit.xml' -f $buildInfo.ModuleName)
-            PassThru   = $true
+            CodeCoverage = @{
+                Enabled    = $true
+                OutputPath = [string](Join-Path -Path $buildInfo.Path.Build.Output -ChildPath pester-codecoverage.xml)
+            }
+            TestResult   = @{
+                Enabled    = $true
+                OutputPath = [string](Join-Path -Path $buildInfo.Path.Build.Output -ChildPath (
+                    '{0}-nunit.xml' -f $buildInfo.ModuleName
+                ))
+            }
+            Output       = @{
+                Verbosity = 'Detailed'
+            }
         }
-        if (Test-Path $buildInfo.Path.Build.RootModule) {
-            $params.Add('CodeCoverage', $buildInfo.Path.Build.RootModule)
-            $params.Add('CodeCoverageOutputFile', (Join-Path $buildInfo.Path.Build.Output 'pester-codecoverage.xml'))
-        }
-        Invoke-Pester @params
+        $pester = Invoke-Pester -Configuration $configuration
     }
 
     if ($buildInfo.BuildSystem -eq 'Desktop') {
